changed type for vendor and product IDs to NSUInteger
git-svn-id: svn://orca.physics.unc.edu/Orca@11725 fd86f90f-2c0a-422a-b972-84106d8db31b

+++ Source/Objects/Hardware/Camac/CCUSB/ORCCUSBModel.h
@@ -109,8 +109,8 @@
 - (void) setUsbInterface:(ORUSBInterface*)anInterface;
 - (NSString*) usbInterfaceDescription;
 - (void) registerWithUSB:(id)usb;
-- (unsigned long) vendorID;
-- (unsigned long) productID;
+- (NSUInteger) vendorID;
+- (NSUInteger) productID;
 - (NSString*) title;
 - (NSString*) settingsLock;
 - (void) checkInterface;
