+++ Source/Objects/Hardware/USB/ADU200/ORADU200Model.h
@@ -62,8 +62,8 @@
 - (void) setUsbInterface:(ORUSBInterface*)anInterface;
 - (NSString*) serialNumber;
 - (void) setSerialNumber:(NSString*)aSerialNumber;
-- (unsigned long) vendorID;
-- (unsigned long) productID;
+- (NSUInteger) vendorID;
+- (NSUInteger) productID;
 - (NSString*) usbInterfaceDescription;
 - (void) setRelayState:(unsigned short)index withValue:(BOOL)aState;
 - (BOOL) relayState:(unsigned short)index;
