+++ Source/Objects/Hardware/USB/MCA927/ORMCA927Model.h
@@ -159,8 +159,8 @@ typedef struct MCA927Registers {
 - (void) setUsbInterface:(ORUSBInterface*)anInterface;
 - (NSString*) serialNumber;
 - (void) setSerialNumber:(NSString*)aSerialNumber;
-- (unsigned long) vendorID;
-- (unsigned long) productID;
+- (NSUInteger) vendorID;
+- (NSUInteger) productID;
 - (NSString*) usbInterfaceDescription;
 - (void) interfaceAdded:(NSNotification*)aNote;
 - (void) interfaceRemoved:(NSNotification*)aNote;
